refactor: 레포지토리 패키지 추가 및 패키지 경로 수정
