+++ app/src/main/java/woowacourse/shopping/AppContainer.kt
@@ -5,11 +5,11 @@ import androidx.room.Room
 import kotlinx.serialization.json.Json
 import okhttp3.HttpUrl.Companion.toHttpUrl
 import okhttp3.OkHttpClient
-import woowacourse.shopping.data.CartRepository
-import woowacourse.shopping.data.CartRepositoryImpl
-import woowacourse.shopping.data.ProductRepository
-import woowacourse.shopping.data.ProductRepositoryImpl
 import woowacourse.shopping.data.local.ShoppingDatabase
+import woowacourse.shopping.data.repository.cart.CartRepository
+import woowacourse.shopping.data.repository.cart.CartRepositoryImpl
+import woowacourse.shopping.data.repository.product.ProductRepository
+import woowacourse.shopping.data.repository.product.ProductRepositoryImpl
 
 object AppContainer {
     private var database: ShoppingDatabase? = null
