+++ app/src/main/java/woowacourse/shopping/data/repository/product/ProductRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.repository.product
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
@@ -6,6 +6,7 @@ import kotlinx.serialization.json.Json
 import okhttp3.HttpUrl
 import okhttp3.OkHttpClient
 import okhttp3.Request
+import woowacourse.shopping.data.network.ProductDto
 import woowacourse.shopping.domain.Product
 
 class ProductRepositoryImpl(
